Add KSampler node, Remove KSampler Mapping from api, Remove Scheduler from AIR props, Better UI
